Fixed coverage

+++ elephant_vending_machine/views.py
@@ -199,7 +199,8 @@ def upload_image():
             try:
                 add_remote_image(save_path, filename)
             except CalledProcessError:
-                os.remove(os.path.join(save_path, filename))
+                if filename in os.listdir(save_path):
+                    os.remove(os.path.join(save_path, filename))
                 response = "Error: Failed to copy file to hosts. ", \
                   "Image not saved, please try again"
                 response_code = 500
